Extract reuseable ModelDescriptionMarkdown from OpenRouter model picker
